WIP: ignition advance chain in tenths of a degree
Moves the ignition timing domain from whole degrees to tenths, so the closed
loop idle advance controller has an actuator resolution finer than 1 degree.
updateIdleAdvanceClosedLoop() already integrated in tenths internally and threw
the precision away in a final cast to int8_t; that cast is now gone.

Done:
 - crankMaths: microsPerTenthDegree / tenthsPerMicro (Q24.8, Q1.15),
   angleTenthsToTimeMicroSec, angleTenthsToTimerTicks,
   timeToAngleTenthsPerMicroSec, ignitionLimitsTenths, degreesToTenths
 - statuses: advance/advance1/advance2 -> int16_t tenths
 - corrections: whole ignition correction chain -> int16_t tenths
 - secondaryTables: constrainAdvance + spark2 blending (widened to int32
   where the tenths product would overflow int16)
 - schedule_calcs: discharge/charge angle + _calculateAngularTimeTenths
 - scheduler_ignition_controller: dwellAngle in tenths

NOT DONE - do not flash this commit:
 - decoders.cpp: ~20 end-teeth thresholds still compare tenths to degrees
 - knock_window.cpp: delayAngle mixes tenths and degrees
 - logger/comms: still emit tenths as if they were degrees
 - spark table half-degree rescale + tune migration
 - tests not yet updated

+++ speeduino/statuses.h
@@ -96,9 +96,9 @@ struct statuses {
   volatile uint16_t actualDwell;    ///< actual dwell time if new ignition mode is used (in uS)
   byte dwellCorrection; /**< The amount of correction being applied to the dwell time (in unit ...). */
   byte battery10;     /**< The current BRV in volts (multiplied by 10. Eg 12.5V = 125) */
-  int8_t advance;     /**< The current advance value being used in the spark calculation. Can be the same as advance1 or advance2, or a calculated value of both */
-  int8_t advance1;    /**< The advance value from ignition table 1 */
-  int8_t advance2;    /**< The advance value from ignition table 2 */
+  int16_t advance;    /**< The current advance being used in the spark calculation, in **tenths of a degree** (@see ANGLE_TENTHS_PER_DEGREE). Can be the same as advance1 or advance2, or a calculated value of both */
+  int16_t advance1;   /**< The advance from ignition table 1, in **tenths of a degree** */
+  int16_t advance2;   /**< The advance from ignition table 2, in **tenths of a degree** */
   uint16_t corrections; /**< The total current corrections % amount */
   uint16_t AEamount;    /**< The amount of acceleration enrichment currently being applied. 100=No change. Varies above 255 */
   uint16_t wallFuel;     /**< Current amount of fuel on intake walls (wall wetting model state) */
